bin {partitions}	R Documentation

Sundry binary functionality

Description

Utilities to convert things to binary

Usage

tobin(n, len, check=TRUE)
todec(bin)
comptobin(comp, check=TRUE)
bintocomp(bin,  use.C=TRUE, check=TRUE)

Arguments

n
Integer, to be converted to binary by function tobin()

len
Length of the binary vector returned by function tobin()

bin
Binary: a vector of 0s and 1s

comp
A composition

check
Boolean, with default TRUE meaning to perform various checks

use.C
Boolean, with default TRUE meaning to use C

Details

These functions are not really intended for the end user; they are used in nextcomposition().

Function tobin() converts integer n to a binary string of length len

Function todec() converts a binary string to decimal, so todec(tobin(n,i))==n, provided i is big enough

Function comptobin() converts a composition to binary

Function bintocomp() converts a binary string to a composition

Author(s)

Robin K. S. Hankin

References

Wikipedia contributors, 2020. “Composition (combinatorics) — Wikipedia, The Free Encyclopedia”, https://en.wikipedia.org/w/index.php?title=Composition_(combinatorics)&oldid=944285378

Examples

tobin(10,5)
todec(tobin(10,5))
comptobin(c(1,1,4))
bintocomp(c(1,1,0,0,1,1,1,1))

[Package partitions version 1.10-1 Index]