print.partition {partitions}	R Documentation

Print methods for partition objects and equivalence objects

Description

A print method for partition objects, summary partition objects, and equivalence classes. Includes various configurable options

Usage

## S3 method for class 'partition'
print(x, mat = getOption("matrixlike"), h = getOption("horiz"), ...)
## S3 method for class 'summary.partition'
print(x, ...)
## S3 method for class 'equivalence'
print(x, sep = getOption("separator"), ...)

Arguments

x
Object to be printed: an object of class either partition or summary.partition

mat
Boolean, with TRUE meaning to print like a matrix, and any other value meaning to print without column names (which usually results in more compact appearance)

h
Boolean governing the orientation of the printed matrix, with TRUE meaning to print with the rows being the partitions and any other value (the default) meaning to print the transpose

sep
Character vector, with special value of NULL interpreted as a comma; see examples section

...
Further arguments provided for compatibility

Author(s)

Robin K. S. Hankin

Examples

print(parts(5))

summary(parts(7))

listParts(3)
options(separator="")
listParts(5)

[Package partitions version 1.10-1 Index]