summary.partition {partitions}	R Documentation

Provides a summary of a partition

Description

Provides a summary of an object of class partition: usually the first and last few partitions (columns)

Usage

## S3 method for class 'partition'
summary(object, ...)

Arguments

object
Partition

...
Further arguments; see details section below

Details

The ellipsis arguments are used to pass how many columns at the start and the end of the matrix are selected; this defaults to 10.

The function is designed to behave as expected: if there is an argument named “n”, then this is used. If there is no such argument, the first one is used.

Value

A summary object is a list, comprising three elements:

shortened
Boolean, with TRUE meaning that the middle section of the matrix is ommitted, and FALSE meaning that the entire matrix is returned because n is too big

n
Number of columns to return at the start and the end of the matrix

out
Matrix returned: just the first and last n columns (if shortened is TRUE), or the whole matrix if not

Author(s)

Robin K. S. Hankin

Examples

summary(parts(7))

summary(parts(11),3)

[Package partitions version 1.10-1 Index]